fix(electron): moved server outside of cli
